% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_MR.R
\name{forest_plot_MR}
\alias{forest_plot_MR}
\title{Calculating the area under the curve after developing the category predictive model}
\usage{
forest_plot_MR(
  file = "",
  column_num = 10,
  shape_col = "#4575b4",
  line_col = "#CE5C69",
  text_col = "#4575b4",
  P_threshold = 0.05,
  OR_range = c(0, 2),
  cutoff = c(0, 1, 2)
)
}
\arguments{
\item{res.by.ML.Dev.Pred.Category.Sig}{Output of function ML.Dev.Pred.Category.Sig}

\item{cohort.for.cal}{A data frame with the 'ID' and 'Var' as the first two columns. Starting in the fourth column are the variables that contain variables of the model you want to build. The second column 'Var' only contains 'Y' or 'N'.}
}
\value{
A data frame containing the AUC of each predictive model.
}
\description{
Calculating the area under the curve after developing the category predictive model
}
