% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_MR.R
\name{result_MR}
\alias{result_MR}
\title{Organize and export the MR results.}
\usage{
result_MR(folder.path = "", exposure.cat = "", outcome.name = "")
}
\arguments{
\item{folder.path}{Directory for storing MR results files.}

\item{exposure.cat}{The category of the exposure factor.}

\item{outcome.name}{The name of the outcome factor.}
}
\description{
Organize and export the results of the Mendelian Randomization (MR) analysis in a clear and structured manner. This involves collecting the key findings, such as effect estimates, confidence intervals, and p-values, and arranging them in a well-labeled table for easy interpretation.
}
\examples{
result_MR (folder.path='MR_result/',
           exposure.cat='Biomarkers'
           outcome.name='PCA')

}
\author{
Junxiao Shen
}
